Added JsonAaCReaderStrategy, so that it can be used to generate Diagrams from JSON AaC
